backend/services/payment-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
python-dotenv==1.0.1
stripe==8.0.0
boto3==1.34.0
